Add first interface object ListGroupProp
Instead of hardcoding the list of city names and the heading,
we can set up variables in an interface object outside of
the ListGroup component function, and then pass each variable,
item and heading, to the function.

These variables can instead be defined in the central App component.
Later we can set these up to be changed dynamically.