Create onSelectItem function prop
This function is a property of ListGroup interface.
It allows us to notify the App component of when an item
in the list has been clicked on and selected.
So instead of defining what should happen on click within the ListGroup
component, we can define this elsewhere.
This means the component can be reused for different purposes in which
a different outcome occurs onSelectItem

We also wrote an event handler function called handleSelectItem in the
App component. For now, all this done is print the name of the item
to the console, so we can check that the event handler works.