Add the Alert component
Commented out the rendering of ListGroup component within App component.
Created a new component called Alert with className alert.
Tested this compoennt and gave it a children prop of type ReactNode.
Used this prop to insert text and HTML elements into the component.
Added notes on props vs states to the tutoring-app-notes.txt file.
Installed the ES7+ extension which enables the use of a shortcut,
rafce, to autogenerate the export and function template for a component.